Move nvl() function to header.js
