+++ pronto/static/header.js
@@ -300,3 +300,10 @@ export function finaliseHeader(signatureAcc) {
             return getCurrentUser();
         });
 }
+
+export function nvl(expr1, expr2, expr3) {
+    if (expr1 !== null && expr1 !== undefined)
+        return expr3 ? expr3 : expr1;
+    else
+        return expr2;
+}
