+++ pronto/static/modules/database_integrated.js
@@ -1,6 +1,5 @@
-import {finaliseHeader} from "../header.js"
+import {finaliseHeader, nvl} from "../header.js"
 import * as ui from "../ui.js";
-import {nvl} from '../utils.js';
 import {checkEntry} from '../events.js';
 import {getSignatureComments, postSignatureComment} from "../comments.js";
