--- pronto/static/utils.js
@@ -1,6 +1,0 @@
-export function nvl(expr1, expr2, expr3) {
-    if (expr1 !== null && expr1 !== undefined)
-        return expr3 ? expr3 : expr1;
-    else
-        return expr2;
-}
